Add a GitHub integration
Raw webhook events from GitHub are now handled, parsed and trigger more
specific events that can be listened for to provide notifications or
trigger builds. (This is currently not used, though.)

Events that are handled include team membership changes (to keep the
trusted users list up-to-date), push, PR, PR review comments, commit
comments, and issue comments.

Hooks are added automatically and "maintained" on repos and the org.

This is inspired by the implementation of Dolphin Central.
Thanks for the ideas :)